Refactoring table
