+++ src/components/OutputViews/TabellenOutput.jsx
@@ -1,10 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import DataTable from "./DataTable";
-import { getPartyColor } from "../../utils/getPartyColor";
-import { Text, Flex } from "@chakra-ui/react";
-import { Icon } from "@chakra-ui/icons"
-
+import ZugriffOutput from './ZugriffOutput'
 
 TabellenOutput.propTypes = {
   tableData: PropTypes.array,
@@ -15,37 +12,7 @@ export default function TabellenOutput({ tableData, assignmentSequence }) {
   let columns = [];
   let data = [];
 
-  function ZugriffOutput({ cell }) {
-    console.log(cell);
-    const value = cell?.value;
-    if (Array.isArray(value)) {
-      return "Mehrdeutig! " + value.join(" oder ");
-    } else {
-      const partyNames = Object.keys(tableData[0].seats);
-      const totalPartyCount = partyNames.length;
-      const index = partyNames.findIndex((elem) => elem == value)
-      const partyColor = getPartyColor(value, index, totalPartyCount)
-
-      const CircleIcon = (props) => (
-        <Icon viewBox="0 0 200 200" {...props}>
-          <path
-            fill="currentColor"
-            d="M 100, 100 m -75, 0 a 75,75 0 1,0 150,0 a 75,75 0 1,0 -150,0"
-          />
-        </Icon>
-      )
-
-      return(
-        <Flex align='center'>
-          <CircleIcon color={partyColor} mr={2} boxSize={12}/>
-          <Text flex={1} textAlign='center'>{value}</Text>
-        </Flex>
-      ) 
-    }
-  }
-  ZugriffOutput.propTypes = {
-    cell: PropTypes.object,
-  };
+  
   if (tableData != null && Array.isArray(tableData) && tableData.length >= 1) {
     // set up columns
     const partyNames = Object.keys(tableData[0].seats);
@@ -68,7 +35,7 @@ export default function TabellenOutput({ tableData, assignmentSequence }) {
       {
         Header: "Zugriff",
         accessor: "seat_goes_to",
-        Cell: ZugriffOutput,
+        Cell: ({ cell }) => ZugriffOutput({cell, tableData}),
       },
     ];
     columns = assignmentCol.concat(columns);
